Update build_and_publish.yml
fix1:The Determine Git tag step uses the awk command to extract the version number from the Git tag. This ensures compatibility across different operating systems.

fix2:The Tag Docker image with version step is conditional (if: steps.get_tag.outputs.TAG != '') to only execute if a valid Git tag is present. It tags the Docker image with the extracted version number and pushes it to ghcr.